Moved broadcast_to_shape tests to unit test structure.
